+++ workspace/solutions/kotlin/lib/src/test/kotlin/s2/s1/s3/s2/s_2132_test.kt
@@ -1,0 +1,16 @@
+package s2.s1.s3.s2
+
+import utils.T
+import kotlin.test.Test
+import s2.s1.s3.s2.s_2132_00.Solution as s_2132_00
+
+class s_2132_test {
+    @Test
+    fun test() {
+        val t = T(
+            this::class,
+            s_2132_00::class,
+        )
+        t.run()
+    }
+}