+++ workspace/solutions/rust/src/lib.rs
@@ -15,4 +15,5 @@
 
 pub mod s0;
 pub mod s1;
+pub mod s2;
 pub mod utils;