+++ workspace/solutions/rust/src/s2/s1/s3/s2/mod.rs
@@ -1,0 +1,4 @@
+pub mod s_2132_00;
+pub mod s_2132_test;
+
+pub type TSolution = fn(Vec<Vec<i32>>, i32, i32) -> bool;