+++ workspace/solutions/rust/src/s2/s1/s3/s2/s_2132_test.rs
@@ -1,0 +1,14 @@
+#[cfg(test)]
+mod tests {
+    use super::super::s_2132_00;
+    use crate::utils::{json::ValueExtension, test::t};
+
+    #[test]
+    fn test() -> () {
+        t(
+            module_path!(), //
+            &vec![s_2132_00::SOLUTION],
+            |solve, input| serde_json::Value::from(solve(input[0].to_i32_matrix().unwrap(), input[1].to_i32().unwrap(), input[2].to_i32().unwrap())),
+        );
+    }
+}